Add 'event' type and 'event_rule' to alarm API
This patch adds 'event' type to the alarm type list.

This patch also adds AlarmEventRule which has two attributes; 'event_type' and
'query'. 'event_type' indicate which type(s) of event to be evaluated for the
alarm. 'query' is a list of conditions combined by AND as the same as
AlarmThresholdRule.

APIImpact
Implements: blueprint event-alarm-evaluator

Change-Id: I4681bf6fa9f01a77122d02dd619753bbb0b982ea